% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDLPicasso.R
\name{Func_CombineDataset}
\alias{Func_CombineDataset}
\title{Title Func_CombineDataset}
\usage{
Func_CombineDataset(
  MyDatasetName = "N10_WMMultiomeKatahdin",
  FirstDataset = "PatientDemography",
  SecondDataset = "SampleInformation"
)
}
\arguments{
\item{MyDatasetName}{DatasetName}

\item{FirstDataset}{PatientDemophgraphy}

\item{SecondDataset}{SampleInformation}
}
\value{
Joinned PatientDemography and SampleInformation
}
\description{
Title Func_CombineDataset
}
\examples{
Func_CombineDataset(MyDatasetName="N18_dbMP", FirstDataset="PatientDemography", SecondDataset="SampleInformation")
}
