% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDLPicasso.R
\name{Func_SummaryPiechart}
\alias{Func_SummaryPiechart}
\title{Title Func_SummaryPiechart}
\usage{
Func_SummaryPiechart(MyDatasetName = "N19_dbMP", InterestDataType = "Race")
}
\arguments{
\item{MyDatasetName}{DatasetName}

\item{InterestDataType}{Datatype of my interest}
}
\value{
Piechart
}
\description{
Title Func_SummaryPiechart
}
\examples{
Func_SummaryPiechart(MyDatasetName="N19_dbMP",  InterestDataType="Race")
}
