% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDLPicasso.R
\name{PCDD_Server}
\alias{PCDD_Server}
\title{Title PCDD_Server}
\usage{
PCDD_Server(input, output)
}
\arguments{
\item{input}{DatasetName}

\item{output}{Piechart}
}
\value{
PCDL plots
}
\description{
Title PCDD_Server
}
\examples{
PCDD_Server(input,output)
}
