% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCDLPicasso.R
\name{Func_ClinicalDataTable}
\alias{Func_ClinicalDataTable}
\title{Title Func_ClinicalDataTable}
\usage{
Func_ClinicalDataTable(
  MyDatasetName = "N10_WMMultiomeKatahdin",
  ClinicalDataType = "Race",
  SubType = "White"
)
}
\arguments{
\item{MyDatasetName}{DatasetName}

\item{ClinicalDataType}{Race or Gender}

\item{SubType}{Subtypes in clinical data type}
}
\value{
CombinedData_BySubType
}
\description{
Title Func_ClinicalDataTable
}
\examples{
Func_ClinicalDataTable(MyDatasetName="N18_dbMP",  ClinicalDataType="Race",  SubType="Asian")
}
